+++ decider/load.go
@@ -245,5 +245,5 @@ func (r *Repository[State]) loadFromSnapshot(
 		return zero, 0, err
 	}
 
-	return state, snap.Version.Add(len(events)), nil
+	return state, snap.Version.Add(uint(len(events))), nil
 }
